--- apps/samradsgatt/components/DummyData/SubscriptionDummyData.ts
@@ -1,300 +1,0 @@
-const SubscriptionDummyData = {
-    cases: [
-        {
-            id: 0,
-            caseNumber: 'Öll mál',
-            caseTitle: 'Tilkynningar um ný mál',
-            checked: false
-        },
-        {
-            id: 1,
-            caseNumber: 'Öll mál',
-            caseTitle: 'Tilkynningar um ný mál, breyttan umsagnarfrest, umsagnarfrest sem er að renna út og birtingu niðurstaðna',
-            checked: false
-        },
-        {
-            id: 2,
-            caseNumber: '26/2023',
-            caseTitle: 'Brottfall úreltra reglugerða á fjármálamarkaði',
-            checked: true
-        },
-        {
-            id: 3,
-            caseNumber: '25/2023',
-            caseTitle: 'Drög að reglugerð um (2.) breytingu á reglugerð nr. 241/2004 um val, geymslu og meðferð lyfja á sjúkrahúsum og öðrum heilbrigðisstofnunum',
-            checked: false
-        },
-        {
-            id: 4,
-            caseNumber: '24/2023',
-            caseTitle: 'Frumvarp til laga um breytingu á lögum um tæknifrjóvgun og notkun kynfrumna og fósturvísa manna til stofnfrumurannsókna, nr. 55/1996 (geymsla og nýting fósturvísa og kynfrumna)',
-            checked: false
-        },
-        {
-            id: 5,
-            caseNumber: '23/2023',
-            caseTitle: 'Áform um breytingu á lögum um ríkislögmann, nr. 51/1985',
-            checked: false
-        },
-        {
-            id: 6,
-            caseNumber: '2/2023',
-            caseTitle: 'Tillaga til þingsályktunar um aðgerðaáætlun gegn hatursorðræðu 2023-2026',
-            checked: false
-        },
-        {
-            id: 7,
-            caseNumber: '22/2023',
-            caseTitle: 'Frumvarp til laga um breytingu á lögum nr. 116/2006, um stjórn fiskveiða (svæðaskipting strandveiða)',
-            checked: false
-        },
-        {
-            id: 8,
-            caseNumber: '20/2023',
-            caseTitle: 'Neyðarbirgðir eldsneytis',
-            checked: false
-        },
-        {
-            id: 9,
-            caseNumber: '15/2023',
-            caseTitle: 'Drög að reglugerð um notkunar mannalyfja af mannúðarástæðum',
-            checked: false
-        },
-        {
-            id: 10,
-            caseNumber: '19/2023',
-            caseTitle: 'Reglugerð um atvinnusjúkdóma',
-            checked: false
-        },
-        {
-            id: 11,
-            caseNumber: '21/2023',
-            caseTitle: 'Drög að frumvarpi til laga um breytingu á lögum um heilbrigðisstarfsmenn, nr. 34/2012 (heimilisofbeldi)',
-            checked: false
-        },
-        {
-            id: 12,
-            caseNumber: '215/2022',
-            caseTitle: 'Reglugerð um íbúakosningar sveitarfélaga',
-            checked: false
-        },
-        {
-            id: 13,
-            caseNumber: '253/2022',
-            caseTitle: 'Drög að þingsályktunartillögu um aðgerðaáætlun um þjónustu við eldra fólk 2023-2027',
-            checked: false
-        },
-        {
-            id: 14,
-            caseNumber: '7/2023',
-            caseTitle: 'Bráðabirgðaniðurstöður starfshópa - Auðlindin okkar - stefna um sjávarútveg',
-            checked: false
-        },
-        {
-            id: 15,
-            caseNumber: '18/2023',
-            caseTitle: 'Frumvarp til laga um breytingu á lögum um réttindi sjúklinga nr. 74/1997',
-            checked: false
-        },
-        {
-            id: 16,
-            caseNumber: '12/2023',
-            caseTitle: 'Drög að reglugerð um útgáfu vottorða, álitsgerða, faglegra yfirlýsinga og skýrslur heilbrigðisstarfsmanna',
-            checked: false
-        },
-        {
-            id: 17,
-            caseNumber: '17/2023',
-            caseTitle: 'Breyting á lögum í tengslum við raforkueftirlit Orkustofnunar',
-            checked: false
-        },
-        {
-            id: 18,
-            caseNumber: '247/2022',
-            caseTitle: 'Drög að þingsályktunartillögu um aðgerðaáætlun í geðheilbrigðismálum 2023-2027',
-            checked: false
-        },
-        {
-            id: 19,
-            caseNumber: '16/2023',
-            caseTitle: 'Drög að reglugerð um breytingu á reglugerð nr. 944/2014, um öryggi leikfanga og markaðssetningu þeirra á Evrópska efnahagssvæðinu',
-            checked: false
-        }
-    ],
-    institutions: [
-        {
-            id: 0,
-            institution: 'Allar stofnanir. Tilkynningar um ný mál',
-            checked: false
-        },
-        {
-            id: 1,
-            institution: 'Allar stofnanir. Tilkynningar um ný mál, breyttan umsagnarfrest, umsagnarfrest sem er að renna út og birtingu niðurstaðna',
-            checked: false
-        },
-        {
-            id: 2,
-            institution: 'Dómsmálaráðuneytið',
-            checked: false
-        },
-        {
-            id: 3,
-            institution: 'Félags- og vinnumarkaðsráðuneytið',
-            checked: false
-        },
-        {
-            id: 4,
-            institution: 'Fjármála- og efnahagsráðuneytið',
-            checked: false
-        },
-        {
-            id: 5,
-            institution: 'Forsætisráðuneytið',
-            checked: false
-        },
-        {
-            id: 6,
-            institution: 'Háskóla-, iðnaðar- og nýsköpunarráðuneytið',
-            checked: false
-        },
-        {
-            id: 7,
-            institution: 'Heilbrigðisráðuneytið',
-            checked: false
-        },
-        {
-            id: 8,
-            institution: 'Innviðaráðuneytið',
-            checked: false
-        },
-        {
-            id: 9,
-            institution: 'Matvælaráðuneytið',
-            checked: false
-        },
-        {
-            id: 10,
-            institution: 'Menningar- og viðskiptaráðuneytið',
-            checked: false
-        },
-        {
-            id: 11,
-            institution: 'Mennta- og barnamálaráðuneytið',
-            checked: false
-        },
-        {
-            id: 12,
-            institution: 'Umhverfis-, orku- og loftslagsráðuneytið',
-            checked: false
-        },
-        {
-            id: 13,
-            institution: 'Utanríkisráðuneytið',
-            checked: false
-        },
-    ],
-    policyAreas: [
-        {
-            id: 0,
-            policyArea: 'Öll málefnasvið. Tilkynningar um ný mál.',
-            checked: false
-        },
-        {
-            id: 1,
-            policyArea: 'Öll málefnasvið. Tilkynningar um ný mál, breyttan umsagnarfrest, umsagnarfrest sem er að renna út og birtingu niðurstaðna.',
-            checked: false
-        },
-        {
-            id: 2,
-            policyArea: 'Almanna- og réttaröryggi',
-            checked: false
-        },
-        {
-            id: 3,
-            policyArea: 'Almennur varasjóður og sértækar fjárráðstafanir',
-            checked: false
-        },
-        {
-            id: 4,
-            policyArea: 'Dómstólar',
-            checked: false
-        },
-        {
-            id: 5,
-            policyArea: 'Ferðaþjónusta',
-            checked: false
-        },
-        {
-            id: 6,
-            policyArea: 'Fjölmiðlun',
-            checked: false
-        },
-        {
-            id: 7,
-            policyArea: 'Fjölskyldumál',
-            checked: false
-        },
-        {
-            id: 8,
-            policyArea: 'Framhaldsskólastig',
-            checked: false
-        },
-        {
-            id: 9,
-            policyArea: 'Hagskýrslugerð og grunnskrár',
-            checked: false
-        },
-        {
-            id: 10,
-            policyArea: 'Háskólastig',
-            checked: false
-        },
-        {
-            id: 11,
-            policyArea: 'Heilbrigðisþjónusta utan sjúkrahúsa',
-            checked: false
-        },
-        {
-            id: 12,
-            policyArea: 'Hjúkrunar- og endurhæfingarþjónusta',
-            checked: false
-        },
-        {
-            id: 13,
-            policyArea: 'Húsnæðis- og skipulagsmál',
-            checked: false
-        },
-        {
-            id: 14,
-            policyArea: 'Landbúnaður',
-            checked: false
-        },
-        {
-            id: 15,
-            policyArea: 'Lyf og lækningavörur',
-            checked: false
-        },
-        {
-            id: 16,
-            policyArea: 'Lýðheilsa og stjórnsýsla velferðarmála',
-            checked: false
-        },
-        {
-            id: 17,
-            policyArea: 'Alþingi og eftirlitsstofnanir þess',
-            checked: false
-        },
-        {
-            id: 18,
-            policyArea: 'Markaðseftirlit og neytendamál',
-            checked: false
-        },
-        {
-            id: 19,
-            policyArea: 'Málefni aldraðra',
-            checked: false
-        },
-    ]
-}
-
-export default SubscriptionDummyData
