+++ apps/samradsgatt/components/Table/SubscriptionTable.css.ts
@@ -1,0 +1,16 @@
+import { theme } from '@island.is/island-ui/theme'
+import { style } from '@vanilla-extract/css'
+
+export const checkmarkIcon = style({
+    verticalAlign: 'bottom'
+})
+
+export const tableRowLeft = style({
+    borderTopLeftRadius: '8px',
+    borderBottomLeftRadius: '8px'
+})
+
+export const tableRowRight = style({
+    borderTopRightRadius: '8px',
+    borderBottomRightRadius: '8px'
+})
