+++ apps/samradsgatt/components/Table/SubscriptionTable.tsx
@@ -1,6 +1,7 @@
-import { Icon, Table as T, Checkbox, Text, Box } from '@island.is/island-ui/core'
-import isEvenCheck from '../../utils/helpers/isEvenCheck'
-import SubscriptionDummyData from '../DummyData/SubscriptionDummyData'
+import React from 'react'
+import { Icon, Table as T, Checkbox, Text } from '@island.is/island-ui/core'
+import * as styles from './SubscriptionTable.css'
+import tableRowBackgroundColor from '../../utils/helpers/tableRowBackgroundColor'
 
 const Headers = {
     cases: ['Málsnr.', 'Heiti máls'],
@@ -8,18 +9,11 @@ const Headers = {
     policyAreas: ['Málefnasvið']
 }
 
-const BackgroundColor = (idx: number) => {
-    const isEven = isEvenCheck(idx)
-    return isEven ? 'blue100' : 'transparent'
-}
-
-const SubscriptionTable = ({ }) => {
+const SubscriptionTable = ({ data, setData }) => {
     const chosenTab = "cases"
-    const dummyData = SubscriptionDummyData[chosenTab]
-
     let headerKey = 0
 
-    const onCheckboxChange = (idx: number) => {
+    const onCheckboxChange = () => {
         
     }
 
@@ -27,19 +21,19 @@ const SubscriptionTable = ({ }) => {
         <T.Table>
             <T.Head>
                 <T.Row>
-                <T.HeadData box={{ background: 'transparent', borderColor: 'transparent' }} key={headerKey++}><Icon icon="checkmark" color="blue400" /></T.HeadData>
+                <T.HeadData box={{ background: 'transparent', borderColor: 'transparent' }} key={headerKey++}><Icon icon="checkmark" color="blue400" className={styles.checkmarkIcon} /></T.HeadData>
                 {Headers[chosenTab].map((header) => (
-                    <T.HeadData box={{ background: 'transparent', borderColor: 'transparent' }} key={headerKey++}>{header}</T.HeadData>
+                    <T.HeadData text={{ variant: 'h4'}} box={{ background: 'transparent', borderColor: 'transparent' }} key={headerKey++}>{header}</T.HeadData>
                 ))}
                 </T.Row>
                 
             </T.Head>
             <T.Body>
-                {dummyData.map((data, idx) => (
-                    <T.Row key={data.id }>
-                        <T.Data borderColor="transparent" box={{ borderRadius: 'standard', background: BackgroundColor(idx)}}><Checkbox checked={data.checked} onChange={() => onCheckboxChange(data.id)} /></T.Data>
-                        <T.Data borderColor="transparent" box={{ background: BackgroundColor(idx)}}><Text variant="h5">{data.caseNumber}</Text></T.Data>
-                        <T.Data borderColor="transparent" box={{ borderRadius: 'standard', background: BackgroundColor(idx)}}><Text variant="h5">{data.caseTitle}</Text></T.Data>
+                {data.map((item, idx) => (
+                    <T.Row key={item.id }>
+                        <T.Data borderColor="transparent" box={{ className: styles.tableRowLeft, background: tableRowBackgroundColor(idx)}}><Checkbox checked={item.checked} onChange={onCheckboxChange} /></T.Data>
+                        <T.Data borderColor="transparent" box={{ background: tableRowBackgroundColor(idx)}}><Text variant="h5">{item.caseNumber}</Text></T.Data>
+                        <T.Data borderColor="transparent" box={{ className: styles.tableRowRight, background: tableRowBackgroundColor(idx)}}><Text variant="medium">{item.caseTitle}</Text></T.Data>
                     </T.Row>
                 ))}
             </T.Body>
