+++ apps/samradsgatt/utils/helpers/tableRowBackgroundColor.ts
@@ -1,0 +1,8 @@
+import isEvenCheck from "./isEvenCheck"
+
+const tableRowBackgroundColor = (idx: number) => {
+    const isEven = isEvenCheck(idx)
+    return isEven ? 'blue100' : 'transparent'
+}
+
+export default tableRowBackgroundColor
